requirements.txt
python-docx
TexSoup
gitpython
ply
bibtexparser
chardet